% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm_methods.R
\name{print.flm}
\alias{print.flm}
\title{Print method for flm objects}
\usage{
\method{print}{flm}(x, ...)
}
\arguments{
\item{x}{An object of class "flm", typically the result of a call to \code{\link{flm}}.}

\item{...}{Additional arguments passed to model printing functions.}
}
\value{
The function is called for printing model information.
}
\description{
Displays a concise summary of a fitted functional linear model object.
}
\details{
Prints basic information about the model including the call, number of observations,
response grid length, and a summary of each predictor.
}
\examples{
\dontrun{
# Fit functional factor regression model
ffr_model <- flm(Y ~ X + w, data = data, K = c(3))

# Print basic model information
print(ffr_model)
}

}
\seealso{
\code{\link{flm}}, \code{\link{summary.flm}}
}
