% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm_methods.R
\name{print.summary.flm}
\alias{print.summary.flm}
\title{Print method for summary.flm objects}
\usage{
\method{print}{summary.flm}(x)
}
\arguments{
\item{x}{An object of class "summary.flm", typically the result of a call to \code{\link{summary.flm}}.}
}
\value{
The function is called for printing summary information.
}
\description{
Displays a formatted summary of a functional linear model.
}
\details{
Presents a detailed overview of model fit statistics, model complexity metrics,
and predictor information. Also provides guidance on how to visualize different
aspects of the model using the \code{\link{plot.summary.flm}} function.
}
\examples{
\dontrun{
# Fit functional factor regression model
ffr_model <- flm(Y ~ X + w, data = data, K = c(3))

# Compute and print summary statistics
model_summary <- summary(ffr_model)
print(model_summary)
}

}
\seealso{
\code{\link{flm}}, \code{\link{summary.flm}}, \code{\link{plot.summary.flm}}
}
